# docs/kazima-ai/openapi.yaml
openapi: 3.1.0
info:
  title: Kazima AI Scholarly Assistant API
  version: 1.0.0
  summary: Source-grounded API for Kazima's cultural and scholarly assistant
servers:
  - url: https://kazima.org
paths:
  /api/kazima-assistant/query:
    post:
      summary: Answer a user query with retrieval, citations, and confidence metadata
      operationId: queryKazimaAssistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantQueryRequest"
      responses:
        "200":
          description: Source-grounded answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantQueryResponse"
  /api/kazima-assistant/retrieve:
    post:
      summary: Retrieve candidate Kazima sources without generation
      operationId: retrieveKazimaSources
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RetrievalRequest"
      responses:
        "200":
          description: Ranked retrieval hits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RetrievalResponse"
...
components:
  schemas:
    AssistantMode:
      type: string
      enum: [brief, research]
    AssistantScope:
      type: string
      enum:
        - kazima_primary
        - kazima_primary_plus_context
        - general_knowledge
        - needs_verification
    ConfidenceLevel:
      type: string
      enum: [high, medium, low]
    SourceContentType:
      type: string
      enum:
        - article
        - biography
        - manuscript
        - page
        - institution
        - timeline
        - topic
    SourceExcerpt:
      type: object
      required: [sourceId, title, type, excerpt, score]
      properties:
        sourceId:
          type: string
        chunkId:
          type: string
        title:
          type: string
        type:
          $ref: "#/components/schemas/SourceContentType"
        excerpt:
          type: string
        url:
          type: string
        score:
          type: number
        sectionTitle:
          type: string
        pageNumber:
          type: integer
    Citation:
      type: object
      required: [id, label, sourceId, excerpt, confidence]
      properties:
        id:
          type: string
        label:
          type: string
        sourceId:
          type: string
        excerpt:
          type: string
        url:
          type: string
        confidence:
          $ref: "#/components/schemas/ConfidenceLevel"
        note:
          type: string
    AnswerSection:
      type: object
      required: [title, body]
      properties:
        title:
          type: string
        body:
          type: string
    AssistantQueryRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
        mode:
          $ref: "#/components/schemas/AssistantMode"
        maxSources:
          type: integer
          minimum: 1
          maximum: 10
        includeReadMore:
          type: boolean
        userIntent:
          type: string
        pageContext:
          type: object
          properties:
            pageId:
              type: integer
            pageTitle:
              type: string
            pageType:
              type: string
    AssistantQueryResponse:
      type: object
      required:
        - query
        - mode
        - scope
        - confidence
        - answer
        - summary
        - sections
        - citations
        - readMore
        - followUpQuestions
        - retrieval
      properties:
        query:
          type: string
        normalizedQuery:
          type: string
        mode:
          $ref: "#/components/schemas/AssistantMode"
        scope:
          $ref: "#/components/schemas/AssistantScope"
        confidence:
          $ref: "#/components/schemas/ConfidenceLevel"
        answer:
          type: string
        summary:
          type: string
        sections:
          type: array
          items:
            $ref: "#/components/schemas/AnswerSection"
        citations:
          type: array
          items:
            $ref: "#/components/schemas/Citation"
        readMore:
          type: array
          items:
            type: object
            required: [title, url]
            properties:
              title:
                type: string
              url:
                type: string
        followUpQuestions:
          type: array
          items:
            type: string
        disclaimers:
          type: array
          items:
            type: string
        retrieval:
          type: object
          required: [returnedSources, sources]
          properties:
            returnedSources:
              type: integer
            totalCandidates:
              type: integer
            sources:
              type: array
              items:
                $ref: "#/components/schemas/SourceExcerpt"
    RetrievalRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
        maxSources:
          type: integer
        filters:
          type: object
          additionalProperties: true
    RetrievalResponse:
      type: object
      required: [query, total, results]
      properties:
        query:
          type: string
        total:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/SourceExcerpt"
...